move banner components into their respective locations
